Refactor metadata.py functionality into setup.py
